requirements.txt
requests>=2.31.0
pyyaml>=6.0
click>=8.1.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
tenacity>=8.2.0
rich>=13.0.0
tqdm>=4.65.0
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
